task3/requirements.txt
pandas
numpy
requests
nltk
python-dotenv
